# Bup: Optimize finding edges