In the case where the input is large enough, there's a lot to be gained
by avoiding flowing all bytes through the window.

On my machine, this leads to a pretty huge speed up:

```
roll/Bup/split          time:   [118.43 us 118.49 us 118.55 us]
                        thrpt:  [1.0297 GiB/s 1.0302 GiB/s 1.0308 GiB/s]
                 change:
                        time:   [-83.205% -83.125% -83.048%] (p = 0.00 < 0.05)
                        thrpt:  [+489.91% +492.60% +495.42%]
                        Performance has improved.
```