@Dr-Emann 2022-07-06
I'm not actually using this for anything. I'm really interested in rolling hashes/CDC though.

I'd like to maybe work on some breaking changes like was suggested in #13, to have a trait for a rolling hash, and a trait for a chunker.